Add signature
Google scholar fails to meet the ethical standards of the research publication industry with regards to name changes. This alone would be enough to stop using the platform, but in my use the search results and bibliographic information provided are often inaccurate and unreliable which only compounds the issue of unethical naming practices for trans scholars. Given their failures as a platform, I will not use Google Scholar until they meet the demands listed in this document.